Enable python syntax highlighting and replace unicode quotes
Markdown code blocks did not contain a programming language near the
triple backtick initializer (```). Adding the word "python" directly
after enabled syntax highlighting, which should make things easier on
the eyes.

As ‘ (U+2018) and ’ (U+2019) are invalid in Python, running the original
code directly will throw a SyntaxError if those characters are
preserved. Replacing them with ' ensures the code can be run (unless
there are other errors present in the examples).

Additional note: the Gemfile.lock was modified, possibly due to running
bundle exec jekyll serve on a linux machine. Unsure if this means that
the Gemfile.lock needs to be put in a .gitignore.